build: fix blazebuild scripts
