+++ blaze/wrapper/blaze_wrapper.mjs
@@ -79,7 +79,7 @@ function getProperty(name, def) {
 
 function determineIndexFile() {
     const blazeSrcPath = getProperty("blaze.srcpath", "build_src");
-    return path.resolve(PROJECT_DIR, blazeSrcPath, "src/main/typescript/index.ts");
+    return path.resolve(PROJECT_DIR, blazeSrcPath, "src/main/typescript/cli.ts");
 }
